+++ CHANGELOG.md
@@ -33,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI: Archived rooms can now be opened in the timeline again
 - SDK: Fix broken os_log logging on iOS
 - SDK: Fix mis-compilation in JavaScript resulting in the 'create profile' button not working
-- UI: Fix timeline not scrolling to / scrolling away from the unread marker
+- UI: Fix various timeline scrolling issues
 - WEB,A11Y: Fix message bubbles not being read out in some browser/screeenreader combinations
 - SDK: Fix NotOnMainThreadException when verifying another user
 - UI: Fix naming of untrusted users